refactor(consumption): extract ConfidenceTierBadge to keep stats card under 400 lines
file_length_test (#1680) flagged consumption_stats_card.dart at 407
lines after the #2027 badge addition. Splits the new badge into
`confidence_tier_badge.dart` (~85 lines). The inline `_ConfidenceTierBadge`
becomes the public `ConfidenceTierBadge`; consumption_stats_card.dart
drops to 332 lines.
